chore: improved typing for evmTokenApi
